clean up types for sample reg
